split pinecone-client.ts with 3 sub files
